add service

+++ java_spring_table_jpa/demo/src/main/java/com/example/demo/controller/UtenteController.java
@@ -1,21 +1,20 @@
 package com.example.demo.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.demo.Utente;
+import com.example.demo.entities.Utente;
 import com.example.demo.repository.UtenteRepo;
+import com.example.demo.service.UtenteService;
 
 
 @RestController
@@ -24,10 +23,13 @@ public class UtenteController {
     @Autowired
     private UtenteRepo utenteRepo;
 
+    @Autowired
+    private UtenteService utenteService;
+
     // chiamaete get, tutti gli utenti
     @GetMapping("/utenti")
     public List<Utente> getUtenti(){
-        return utenteRepo.findAll();
+        return utenteService.getAllUtenti();
     }
 
 
@@ -63,9 +65,16 @@ public class UtenteController {
 
     @DeleteMapping("/utente/{UtenteId}")
     public Utente deleteById(@PathVariable  Integer UtenteId) {
-        Utente pippo = utenteRepo.findById(UtenteId).get();
-        utenteRepo.deleteById(UtenteId);
-        return pippo;
+        Optional<Utente> verifyPresent = utenteRepo.findById(UtenteId);
+        if (!verifyPresent.isPresent()) {
+            Utente UtenteNull = new Utente(UtenteId);
+            return UtenteNull;
+        }else{
+            Utente utenteDeleted = utenteRepo.findById(UtenteId).get();
+            utenteRepo.deleteById(UtenteId);
+            return utenteDeleted;
+        }
+        
     }
